+++ plugins/LH_ColorSwatch/LH_ColorSwatch.cpp
@@ -44,8 +44,8 @@ lh_class *LH_ColorSwatch::classInfo()
     return &classInfo;
 }
 
-LH_ColorSwatch::LH_ColorSwatch(float lowVal, QColor lowCol,
-               float highVal, QColor highCol,
+LH_ColorSwatch::LH_ColorSwatch(qreal lowVal, QColor lowCol,
+               qreal highVal, QColor highCol,
                bool lockDown) :
     LH_Rectangle(),
     lowVal_(lowVal),
@@ -81,13 +81,13 @@ const char *LH_ColorSwatch::userInit()
     setup_horizontal_->setFlags(LH_FLAG_READONLY | LH_FLAG_HIDDEN | LH_FLAG_NOSAVE_DATA | LH_FLAG_NOSAVE_LINK | LH_FLAG_NOSOURCE | LH_FLAG_NOSINK);
     setup_horizontal_->setValue(false);
 
-    setup_lowerbound_value_ = new LH_Qt_float(this,"Lower Bound Value", lowVal_, std::numeric_limits<float>::min(), std::numeric_limits<float>::max(), LH_FLAG_AUTORENDER);
+    setup_lowerbound_value_ = new LH_Qt_float(this,"Lower Bound Value", lowVal_, std::numeric_limits<qreal>::min(), std::numeric_limits<qreal>::max(), LH_FLAG_AUTORENDER);
     setup_lowerbound_color_ = new LH_Qt_QColor(this, "Lower Bound Color", lowCol_, LH_FLAG_AUTORENDER);
 
-    setup_upperbound_value_ = new LH_Qt_float(this,"Upper Bound Value", highVal_, std::numeric_limits<float>::min(), std::numeric_limits<float>::max(), LH_FLAG_AUTORENDER);
+    setup_upperbound_value_ = new LH_Qt_float(this,"Upper Bound Value", highVal_, std::numeric_limits<qreal>::min(), std::numeric_limits<qreal>::max(), LH_FLAG_AUTORENDER);
     setup_upperbound_color_ = new LH_Qt_QColor(this, "Upper Bound Color", highCol_, LH_FLAG_AUTORENDER);
 
-    setup_current_value_ = new LH_Qt_float(this,"Current Value", lowVal_, std::numeric_limits<float>::min(), std::numeric_limits<float>::max(), LH_FLAG_AUTORENDER);
+    setup_current_value_ = new LH_Qt_float(this,"Current Value", lowVal_, std::numeric_limits<qreal>::min(), std::numeric_limits<qreal>::max(), LH_FLAG_AUTORENDER);
     setup_current_color_ = new LH_Qt_QColor(this, "Current Color", lowCol_, LH_FLAG_AUTORENDER | LH_FLAG_READONLY | LH_FLAG_NOSAVE_DATA | LH_FLAG_NOSAVE_LINK | LH_FLAG_NOSINK );
 
     if(lockDown_)
